Add files via upload
Here, presenting the QuizzApp not only for the general knowledge but also for other domains. Download and explore.
If you're often found bored then you arrived at the right spot, Test your knowledge in different categories and sharpen your skill with the best application the QuizzApp. Which will surely fire up your mental synapses with not only maths, science, and general knowledge but with the plethora of things inside it download and explore? 
Link: https://play.google.com/store/apps/details?id=com.developer.quizapp&hl=en